Progress on Day 9, part 1
I have almost everything updated, but I have not attempted to do
anything at all with the very large numbers that we are supposed to now
be able to handle. Those are going to be...vexing. Since 2^64 is
approximately 10^18 = 18 digits, all of the examples listed look like
they should work, so I'm going to swap to 64-bit integers and hope that
works. :/